@@ -175,25 +175,25 @@ func (ft *OSMFileTest) downloadTestOSMFile() {
 	if _, err := os.Stat(ft.FileName); os.IsNotExist(err) {
 		out, err := os.Create(ft.FileName)
 		if err != nil {
-			ft.T.Fatal(err)
+			ft.Fatal(err)
 		}
 		defer out.Close()
 
 		resp, err := http.Get(ft.FileURL)
 		if err != nil {
-			ft.T.Fatal(err)
+			ft.Fatal(err)
 		}
 		defer resp.Body.Close()
 
 		if resp.StatusCode != http.StatusOK {
-			ft.T.Fatalf("test status code invalid: %v", resp.StatusCode)
+			ft.Fatalf("test status code invalid: %v", resp.StatusCode)
 		}
 
 		if _, err := io.Copy(out, resp.Body); err != nil {
-			ft.T.Fatal(err)
+			ft.Fatal(err)
 		}
 	} else if err != nil {
-		ft.T.Fatal(err)
+		ft.Fatal(err)
 	}
 }
 
@@ -202,21 +202,20 @@ func (ft *OSMFileTest) testDecode() {
 
 	f, err := os.Open(ft.FileName)
 	if err != nil {
-		ft.T.Fatal(err)
+		ft.Fatal(err)
 	}
 	defer f.Close()
 
 	d := newDecoder(context.Background(), &Scanner{}, f)
 	err = d.Start(runtime.GOMAXPROCS(-1))
 	if err != nil {
-		ft.T.Fatal(err)
+		ft.Fatal(err)
 	}
 
 	var n *osm.Node
 	var w *osm.Way
 	var r *osm.Relation
 	var nc, wc, rc uint64
-	var id string
 	idsOrder := make([]string, 0, len(IDsExpectedOrder))
 	for {
 		e, err := d.Next()
@@ -224,7 +223,7 @@ func (ft *OSMFileTest) testDecode() {
 		if err == io.EOF {
 			break
 		} else if err != nil {
-			ft.T.Fatal(err)
+			ft.Fatal(err)
 		}
 
 		switch v := e.(type) {
@@ -233,7 +232,7 @@ func (ft *OSMFileTest) testDecode() {
 			if v.ID == ft.ExpNode.ID {
 				n = v
 			}
-			id = fmt.Sprintf("node/%d", v.ID)
+			id := fmt.Sprintf("node/%d", v.ID)
 			if _, ok := IDs[id]; ok {
 				idsOrder = append(idsOrder, id)
 			}
@@ -242,7 +241,7 @@ func (ft *OSMFileTest) testDecode() {
 			if v.ID == ft.ExpWay.ID {
 				w = v
 			}
-			id = fmt.Sprintf("way/%d", v.ID)
+			id := fmt.Sprintf("way/%d", v.ID)
 			if _, ok := IDs[id]; ok {
 				idsOrder = append(idsOrder, id)
 			}
@@ -251,7 +250,7 @@ func (ft *OSMFileTest) testDecode() {
 			if v.ID == ft.ExpRel.ID {
 				r = v
 			}
-			id = fmt.Sprintf("relation/%d", v.ID)
+			id := fmt.Sprintf("relation/%d", v.ID)
 			if _, ok := IDs[id]; ok {
 				idsOrder = append(idsOrder, id)
 			}
@@ -260,21 +259,21 @@ func (ft *OSMFileTest) testDecode() {
 	d.Close()
 
 	if !reflect.DeepEqual(ft.ExpNode, n) {
-		ft.T.Errorf("\nExpected: %#v\nActual:   %#v", ft.ExpNode, n)
+		ft.Errorf("\nExpected: %#v\nActual:   %#v", ft.ExpNode, n)
 	}
 	roundCoordinates(w)
 	if !reflect.DeepEqual(ft.ExpWay, w) {
-		ft.T.Errorf("\nExpected: %#v\nActual:   %#v", ft.ExpWay, w)
+		ft.Errorf("\nExpected: %#v\nActual:   %#v", ft.ExpWay, w)
 	}
 	if !reflect.DeepEqual(ft.ExpRel, r) {
-		ft.T.Errorf("\nExpected: %#v\nActual:   %#v", ft.ExpRel, r)
+		ft.Errorf("\nExpected: %#v\nActual:   %#v", ft.ExpRel, r)
 	}
 	if ft.ExpNodeCount != nc || ft.ExpWayCount != wc || ft.ExpRelCount != rc {
-		ft.T.Errorf("\nExpected %7d nodes, %7d ways, %7d relations\nGot %7d nodes, %7d ways, %7d relations.",
+		ft.Errorf("\nExpected %7d nodes, %7d ways, %7d relations\nGot %7d nodes, %7d ways, %7d relations.",
 			ft.ExpNodeCount, ft.ExpWayCount, ft.ExpRelCount, nc, wc, rc)
 	}
 	if !reflect.DeepEqual(ft.IDsExpOrder, idsOrder) {
-		ft.T.Errorf("\nExpected: %v\nGot:      %v", ft.IDsExpOrder, idsOrder)
+		ft.Errorf("\nExpected: %v\nGot:      %v", ft.IDsExpOrder, idsOrder)
 	}
 }
 
